+++ CHANGELOG.rst
@@ -16,7 +16,7 @@ Changes
 
 0.5.14 (2021-09-14)
 ==================
-
+* Added missing auto interval properties in Template
 * Added colour overrides to pie chart panel
 * Added missing attributes from xAxis class
 * Added transformations for the Panel class (https://grafana.com/docs/grafana/next/panels/transformations/types-options/#transformation-types-and-options)